Create edit page and pre-fill form data

+++ app/_lib/actions.js
@@ -43,6 +43,8 @@ export async function deleteReservation(bookingId) {
     const guestBookings = await getBookings(session.user.guestId)
     const guestBookingsIds = guestBookings.map((booking) => booking.id);
 
+    if (!guestBookingsIds.includes(bookingId)) throw new Error("You are not allowed to delete this booking")
+
     const { error } = await supabase.from('bookings').delete().eq('id', bookingId);
 
     if (error)
@@ -50,6 +52,11 @@ export async function deleteReservation(bookingId) {
         throw new Error('Booking could not be deleted');
 
     revalidatePath('/account/reservations')
+}
+
+export async function updateBooking(formData) {
+    console.log(formData);
+
 }
 
 export async function signInAction() {
